refactor code & add tests
move functions around where they best fit
